requirements.txt
absl-py==0.13.0
asgiref==3.5.2
astor==0.8.1
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.6.0
bleach==5.0.1
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
Django==3.2.14
django-braces==1.15.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-model-utils==4.2.0
django-star-ratings==0.9.2
django-summernote==0.8.20.0
djangorestframework==3.13.1
dload==0.6
et-xmlfile==1.0.1
fire==0.3.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
icon-font-to-png==0.4.1
idna==2.10
imageio==2.9.0
importlib-metadata==4.6.3
ipython==7.25.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.18.0
joblib==1.0.0
JPype1==1.2.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
konlpy==0.5.2
lxml==4.6.2
Markdown==3.3.4
matplotlib==3.3.3
matplotlib-inline==0.1.2
mglearn==0.1.9
mysqlclient==1.4.6
#@ file:///C:/users/geejun/downloads/mysqlclient-1.4.6-cp37-cp37m-win_amd64.whl
nltk==3.5
numpy==1.20.1
oauthlib==3.1.0
openpyxl==3.0.5
opt-einsum==3.3.0
palettable==3.3.0
pandas==1.1.4
parso==0.8.2
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.19
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.9.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.2
seaborn==0.11.1
selenium==3.141.0
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
sqlparse==0.4.2
statsmodels==0.12.2
stylecloud==0.5.1
sudo==1.0.0
swapper==1.3.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.1.0
tinycss==0.4
tqdm==4.58.0
traitlets==5.0.5
tweepy==3.10.0
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==6.2.0
wordcloud==1.8.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.5.0
